move config related to config module
