--- v0.1/config/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "config"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1"
