+++ v0.1/gateway/Cargo.toml
@@ -17,5 +17,5 @@ native-tls = "0.2"
 http = "1.0.0"
 
 futures-util = { version = "0.3", default-features = false }
-
-config = { path = "../config" }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
