+++ v0.1/gateway/src/main.rs
@@ -1,6 +1,4 @@
-use std::collections;
 use std::env;
-use std::fmt;
 use std::path;
 use std::sync::Arc;
 
@@ -10,62 +8,10 @@ use native_tls::{Identity};
 use tokio::fs;
 use tokio::net::TcpListener;
 
+mod config;
 mod responses;
 
-use config;
 
-
-pub enum ConfigParseError<'a> {
-	HeaderError(http::header::InvalidHeaderValue),
-	UriError(<http::Uri as TryFrom<String>>::Error),
-	Error(&'a str),
-}
-
-impl fmt::Display for ConfigParseError<'_>  {
-  fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-  	match self {
-  		ConfigParseError::HeaderError(io_error) => write!(f, "{}", io_error),
-  		ConfigParseError::UriError(json_error) => write!(f, "{}", json_error),
-  		ConfigParseError::Error(error) => write!(f, "{}", error),
-  	}
-  }
-}
-
-/*
-	create_address_map iterates config.addresses and creates a map of
-	destination URIs indexed by a URI host.
-	ie: Map<example.com, http://some_address:6789>
-	
-	If a URI fails to parse, the entire operation fails
-*/
-fn create_address_map(
-	config: &config::Config,
-) -> Result<
-	collections::HashMap::<String, http::Uri>,
-	ConfigParseError,
-> {
-  let mut hashmap: collections::HashMap::<String, http::Uri> = collections::HashMap::new();
-  for (index, value) in config.addresses.iter() {
-  	let index_uri = match http::Uri::try_from(index) {
-  		Ok(uri) => uri,
-  		Err(e) => return Err(ConfigParseError::UriError(e)),
-  	};
-  	
-  	let host = match index_uri.host() {
-			Some(uri) => uri,
-			_ => return Err(ConfigParseError::Error("could not find host from addresses")),
-		};
-		
-  	let dest_uri = match http::Uri::try_from(value) {
-  		Ok(uri) => uri,
-  		Err(e) => return Err(ConfigParseError::UriError(e)),
-  	};
-  	
-  	hashmap.insert(host.to_string(), dest_uri);
-  }
-  
-  Ok(hashmap)
-}
 
 #[tokio::main]
 async fn main() {
@@ -74,14 +20,14 @@ async fn main() {
       Some(a) => path::PathBuf::from(a),
       None => return println!("argument error:\nconfig params not found."),
   };
-  let config = match config::Config::from_filepath(&args) {
+  let config = match config::from_filepath(&args) {
       Ok(c) => c,
       Err(e) => return println!("configuration error:\n{}", e),
   };
 
 	// get addresses
   let host_address = config.host.clone() + ":" + &config.port.to_string();
-  let addresses = match create_address_map(&config) {
+  let addresses = match config::create_address_map(&config) {
   	Ok(addrs) => addrs,
   	Err(e) => return println!("address map error:\n{}", e),
   };
@@ -90,11 +36,11 @@ async fn main() {
   // tls cert and keys
   let cert = match fs::read(&config.cert_filepath).await {
   	Ok(f) => f,
-  	Err(e) => return println!("file error:\n{}", e),
+  	Err(e) => return println!("cert error:\n{}", e),
   };
   let key = match fs::read(&config.key_filepath).await {
   	Ok(f) => f,
-  	Err(e) => return println!("file error:\n{}", e),
+  	Err(e) => return println!("cert error:\n{}", e),
   };
   let pkcs8 = match Identity::from_pkcs8(&cert, &key) {
   	Ok(pk) => pk,
@@ -109,7 +55,7 @@ async fn main() {
 	let tls_acceptor = match native_tls::TlsAcceptor::builder(pkcs8)
 		.build() {
 			Ok(native_acceptor) => tokio_native_tls::TlsAcceptor::from(native_acceptor),
-			Err(e) => return println!("native_acceptor:\n{}", e),
+			Err(e) => return println!("native_acceptor error:\n{}", e),
 	};
 
 	// sever loop
